change file name to splitter.py
