+++ pyds/primitives/__init__.py
@@ -10,7 +10,7 @@ from .evaluate import Evaluate
 from .featurize import Featurize
 from .infer import Infer
 from .train import Train
-from .tvt_splitter import TVTSplit
+from .splitter import TVTSplit
 
 __all__ = [
     "Primitive",
